refactor: migrate platform-independent code from firmware to core crate (T-nazbq)
Move ~1,600 lines of pure algorithms and types from crates/firmware to
crates/core, replacing firmware copies with re-exports. Three phases:

Phase 1 - Delete duplicate code: Remove firmware copies of arming errors,
parameter block/CRC, differential drive kinematics, and geo functions
that already existed in core.

Phase 2 - Migrate pure algorithms: Move DCM algorithm and geographic
calculations (haversine, bearing, distance) to core with full test suites.

Phase 3 - Extract mixed modules: Separate platform-independent logic from
Embassy/platform wrappers for AHRS traits, path recorder, navigation
controller, heading source, and mission state. Firmware retains only
EmbassyState globals and platform-specific implementations.

Key design decisions:
- NavigationController trait accepts raw f32 lat/lon instead of GpsPosition
- AhrsError uses as_str() for defmt-compatible logging in no_std core
- From<&Waypoint> for PositionTarget moved to core (both types in core)
- Zero #[cfg(feature)] directives in core (enforced by check-core-no-cfg.sh)

All 209 tests pass, RP2350 embedded build succeeds.
