+++ crates/firmware/src/core/parameters/mod.rs
@@ -2,16 +2,18 @@
 //!
 //! This module provides parameter storage and management for the autopilot system.
 //! Parameters are stored in Flash memory with redundant block rotation for wear leveling.
+//!
+//! Core parameter types (block, crc) are provided by `pico_trail_core::parameters`.
 
-pub mod block;
-pub mod crc;
 pub mod registry;
 pub mod saver;
 pub mod storage;
 
-// Re-export commonly used types
-pub use block::{hash_param_name, Parameter, ParameterBlockHeader, ParameterFlags};
-pub use crc::{calculate_crc32, validate_crc32};
+// Re-export core types
+pub use pico_trail_core::parameters::{
+    calculate_crc32, hash_param_name, validate_crc32, Parameter, ParameterBlockHeader,
+    ParameterFlags, MAX_PARAMS,
+};
 pub use registry::{ParamMetadata, ParamType, ParamValue, ParameterRegistry, RegistryError};
 pub use saver::{ParamSaver, SaveRequest};
 pub use storage::{FlashParamStorage, StorageStats};
