+++ crates/firmware/src/rover/mode/auto.rs
@@ -216,7 +216,9 @@ impl<'a> Mode for AutoMode<'a> {
         state.navigation_active = true;
 
         // Navigate to target
-        let output = self.nav_controller.update(&gps, &target, heading, dt);
+        let output = self
+            .nav_controller
+            .update(gps.latitude, gps.longitude, &target, heading, dt);
 
         // Check for waypoint arrival
         if output.at_target {
