+++ crates/firmware/src/rover/mode/circle.rs
@@ -295,7 +295,9 @@ impl<'a> Mode for CircleMode<'a> {
             .ok_or("Failed to calculate target")?;
 
         // Delegate to navigation controller
-        let output = self.nav_controller.update(&gps, &target, heading, dt);
+        let output = self
+            .nav_controller
+            .update(gps.latitude, gps.longitude, &target, heading, dt);
 
         // Apply commands to actuators
         self.actuators.set_steering(output.steering)?;
