+++ crates/firmware/src/rover/mode/guided.rs
@@ -175,7 +175,9 @@ impl<'a> Mode for GuidedMode<'a> {
         state.navigation_active = true;
 
         // Navigate to target
-        let output = self.nav_controller.update(&gps, &target, heading, dt);
+        let output = self
+            .nav_controller
+            .update(gps.latitude, gps.longitude, &target, heading, dt);
 
         // Check for arrival
         if output.at_target {
