+++ crates/firmware/src/rover/mode/smartrtl.rs
@@ -217,7 +217,9 @@ impl<'a> Mode for SmartRtlMode<'a> {
         let target = PositionTarget::new(waypoint.latitude, waypoint.longitude);
 
         // Navigate to waypoint
-        let output = self.nav_controller.update(&gps, &target, heading, dt);
+        let output = self
+            .nav_controller
+            .update(gps.latitude, gps.longitude, &target, heading, dt);
 
         // Check for waypoint arrival
         if output.at_target {
